Local/requirements.txt
strawberry-graphql~=0.256.1
PyYAML~=6.0
rosetta-ce~=1.7.3
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
defusedxml==0.7.1
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.4
fastapi-utils==0.2.1
filelock==3.15.4
funcy==2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
olefile==0.47
orjson==3.10.3
pip==24.2
psutil==5.9.8
pydantic==1.10.18
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.38.5
tldextract==5.1.2
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.44.0
ariadne==0.24.0